repo/requirements.txt
amqp==5.0.2
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
awscli==1.18.211
billiard==3.6.3.0
black==20.8b1
boto3==1.16.51
botocore==1.19.51
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.5
django-redis==5.0.0
django-appconf==1.0.4
django-imagekit==4.0.2
django-storages==1.11.1
djangorestframework==3.12.2
docutils==0.15.2
drf-yasg==1.20.0
et-xmlfile==1.0.1
gevent==21.1.2
greenlet==1.0.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
isort==5.7.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
kombu==5.0.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
openapi-codec==1.3.2
packaging==20.9
pathspec==0.8.1
pilkit==2.0
Pillow==8.2.0
prompt-toolkit==3.0.10
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.0
pylint==2.6.0
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.1
rsa==4.5
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
simplejson==3.17.2
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
watchtower==1.0.4
wcwidth==0.2.5
wrapt==1.12.1
zope.event==4.5.0
django-cors-headers==3.5.0
zope.interface==5.2.0
argon2-cffi==20.1.0
firebase-admin==5.3.0
celery==5.0.5
haversine==2.7.0